--- a/conf/presto/jvm.config
+++ b/conf/presto/jvm.config
@@ -19,8 +19,7 @@
 
 -server
 -Xmx16G
--XX:+UseG1GC
--XX:G1HeapRegionSize=32M
+-XX:+UseZGC
 -XX:+UseGCOverheadLimit
 -XX:+ExplicitGCInvokesConcurrent
 -XX:+HeapDumpOnOutOfMemoryError