--- a/tests/docker-images/latest-version-image/conf/presto/jvm.config
+++ b/tests/docker-images/latest-version-image/conf/presto/jvm.config
@@ -20,8 +20,7 @@
 -server
 -Xms128M
 -Xmx1500M
--XX:+UseG1GC
--XX:G1HeapRegionSize=32M
+-XX:+UseZGC
 -XX:+UseGCOverheadLimit
 -XX:+ExplicitGCInvokesConcurrent
 -XX:+HeapDumpOnOutOfMemoryError