@lhotari 2022-05-25
some tests seem to be hitting https://bugs.openjdk.java.net/browse/JDK-8257534 with ZGC.

[example failure](https://github.com/apache/pulsar/runs/6581941316?check_suite_focus=true#step:10:3409)
```
  Error:  Tests run: 4, Failures: 1, Errors: 0, Skipped: 3, Time elapsed: 0.17 s <<< FAILURE! - in org.apache.pulsar.client.impl.ClientCnxTest
  Error:  testClientCnxTimeout(org.apache.pulsar.client.impl.ClientCnxTest)  Time elapsed: 0.022 s  <<< FAILURE!
  java.lang.NoClassDefFoundError: Could not initialize class org.apache.pulsar.common.protocol.Commands
  	at org.apache.pulsar.client.impl.ClientCnx.<init>(ClientCnx.java:209)
  	at org.apache.pulsar.client.impl.ClientCnxTest.testClientCnxTimeout(ClientCnxTest.java:55)
  	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
  	at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:77)
  	at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
  	at java.base/java.lang.reflect.Method.invoke(Method.java:568)
  	at org.testng.internal.MethodInvocationHelper.invokeMethod(MethodInvocationHelper.java:132)
  	at org.testng.internal.InvokeMethodRunnable.runOne(InvokeMethodRunnable.java:45)
  	at org.testng.internal.InvokeMethodRunnable.call(InvokeMethodRunnable.java:73)
  	at org.testng.internal.InvokeMethodRunnable.call(InvokeMethodRunnable.java:11)
  	at java.base/java.util.concurrent.FutureTask.run(FutureTask.java:264)
  	at java.base/java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1136)
  	at java.base/java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:635)
  	at java.base/java.lang.Thread.run(Thread.java:833)
```

"It may be that the initialization failed due to OOME but is being reported as a NoClassDefFoundError. " ([comment link](https://bugs.openjdk.java.net/browse/JDK-8257534?focusedCommentId=14424036&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-14424036))

@nicoloboschi has been fixing some memory leaks in the tests. There are open PRs #15513 and #15638.