@heesung-sn 2022-05-25
> Would you please describe the OpenMessaging Benchmark setup you used. Driver files, workload, and `terraform.tfvars`. I would like to confirm your results.

Hi, [dave2wave](https://github.com/dave2wave), 

Basically, I used the followings for the test. I see some trivial setup failures from the current OpenMessaging Benchmark tool, so I locally made some changes to make it work. I probably need to raise a cr to update the OpenMessaging Benchmark tool.

```
# workload:  1-topic-100-partitions-1kb-4p-4c-2000k, 100-partitions-1kb-4p-4c-500k
# pulsar version: branch-2.9

# Garbage collection options

# g1gc
PULSAR_GC=" -XX:+UseG1GC -XX:MaxGCPauseMillis=10 -XX:+ParallelRefProcEnabled -XX:+UnlockExperimentalVMOptions -XX:+DoEscapeAnalysis -XX:ParallelGCThreads=12 -XX:ConcGCThreads=12
 -XX:G1NewSizePercent=50 -XX:+DisableExplicitGC -XX:-ResizePLAB"
PULSAR_GC="${PULSAR_GC} -XX:+PerfDisableSharedMem -XX:+AlwaysPreTouch -XX:-UseBiasedLocking"

# zgc
#PULSAR_GC="-XX:+UseZGC -XX:+PerfDisableSharedMem -XX:+AlwaysPreTouch"

# ShenandoahGC
#PULSAR_GC="-XX:+UseShenandoahGC -XX:+PerfDisableSharedMem -XX:+AlwaysPreTouch"

 # Extra options to be passed to the jvm
-PULSAR_EXTRA_OPTS="${PULSAR_EXTRA_OPTS} ${PULSAR_MEM} ${PULSAR_GC} -Dio.netty.leakDetectionLevel=disabled -Dio.netty.recycler.maxCapacity.default=1000 -Dio.netty.recycler.linkCa
pacity=1024"
+PULSAR_EXTRA_OPTS="${PULSAR_EXTRA_OPTS} -Dio.netty.leakDetectionLevel=disabled -Dio.netty.recycler.maxCapacity.default=1000 -Dio.netty.recycler.linkCapacity=1024"

```