@lhotari 2022-06-02
> I also see the above oom from one of the attempts in [example failure](https://github.com/apache/pulsar/runs/6581941316?check_suite_focus=true#step:10:3409). It appears that G1GC is more stable to run the Pulsar unit tests. Do we think we should use G1GC for the unit tests?

@heesung-sn It should be fine to use ZGC also for unit tests. There has been a bad test that has caused those issues. It will be fixed by #15911 .