@heesung-sn 2022-06-03
We also need to review this PR, https://github.com/apache/pulsar/pull/15692 as ZGC will expose this Jvm metrics bug.